# docs/swagger.yaml
basePath: /
definitions:
  delivery.UpdateStateRequest:
    properties:
      new_state:
        type: string
    type: object
...
  models.RouteTransfer:
    properties:
      capacity:
        type: integer
      description:
        type: string
      end_station:
        type: string
      end_time:
        type: string
      id:
        type: integer
      image_uuid:
        type: string
      interval_minutes:
        type: integer
      name:
        type: string
      start_station:
        type: string
      start_time:
        type: string
    type: object
  models.TicketTransfer:
    properties:
      create_time:
        type: integer
      creator_id:
        type: integer
      end_time:
        type: integer
      form_time:
        type: integer
      id:
        type: integer
      moderator_id:
        type: integer
      routes:
        items:
          $ref: '#/definitions/models.RouteTransfer'
        type: array
      state:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yarik_tri@mail.ru
    name: Rip API
  description: Rip
  title: Rip
  version: 1.0.1
paths:
...
  /tickets:
    get:
      description: Get all not draft tickets
      produces:
      - application/json
      responses:
        "200":
          description: Got tickets
          schema:
            items:
              $ref: '#/definitions/models.TicketTransfer'
            type: array
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: List tickets
      tags:
      - Tickets
  /tickets/{ticketID}:
    delete:
      description: Delete ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: ticketID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ticket deleted
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Delete ticket
      tags:
      - Tickets
    get:
      description: Get ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: ticketID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Got ticket
          schema:
            $ref: '#/definitions/models.TicketTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Get ticket
      tags:
      - Tickets
  /tickets/{ticketID}/moderate:
    put:
      consumes:
      - application/json
      description: Moderate formed ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: ticketID
        required: true
        type: integer
      - description: Ticket new state
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/delivery.UpdateStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ticket moderated
          schema:
            $ref: '#/definitions/models.TicketTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Moderate ticket
      tags:
      - Tickets
  /tickets/draft:
    delete:
      description: Delete ticket draft
      responses:
        "200":
          description: Draft deleted
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Delete ticket draft
      tags:
      - Tickets
    get:
      description: Get ticket draft
      produces:
      - application/json
      responses:
        "200":
          description: Ticket draft for current user
          schema:
            $ref: '#/definitions/models.TicketTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Get ticket draft
      tags:
      - Tickets
  /tickets/draft/form:
    put:
      description: Form ticket draft by ID
      parameters:
      - description: Ticket ID
        in: path
        name: ticketID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ticket formed
          schema:
            $ref: '#/definitions/models.TicketTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Form ticket
      tags:
      - Tickets
  /tickets/routes/{routeID}:
    delete:
      description: Delete route from ticket draft
      parameters:
      - description: Ticket ID
        in: path
        name: ticketID
        required: true
        type: integer
      - description: Route ID
        in: path
        name: routeID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Route deleted from ticket draft
          schema:
            $ref: '#/definitions/models.TicketTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Delete route from ticket
      tags:
      - Routes
    post:
      description: Add route to ticket draft by ID
      parameters:
      - description: Route ID
        in: path
        name: routeID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Route added
          schema:
            $ref: '#/definitions/models.TicketTransfer'
        "400":
          description: Incorrect input
          schema: {}
        "500":
          description: Server error
          schema: {}
      summary: Add route to ticket
      tags:
      - Routes
schemes:
- https
- http
swagger: "2.0"